Extract types and utility functions from overview.tab.tsx into dedicated modules
